+++ src/Model/SprintReport/SprintReportProject.php
@@ -6,10 +6,4 @@ class SprintReportProject
 {
     public string $id;
     public string $name;
-
-    public function __construct(string $id, string $name)
-    {
-        $this->id = $id;
-        $this->name = $name;
-    }
 }