+++ src/Model/SprintReport/SprintReportVersion.php
@@ -5,11 +5,6 @@ namespace App\Model\SprintReport;
 class SprintReportVersion
 {
     public string $id;
-    public string $headline;
-
-    public function __construct(string $id, string $headline)
-    {
-        $this->id = $id;
-        $this->headline = $headline;
-    }
+    public string $name;
+    public string $projectTrackerId;
 }