+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Added commands to manage data providers.
 * Changed how errors are handled in Leantime api calls.
 * Modified getSprintReportData to work with Leantime data
 * Added project lead to client when syncing projects.
@@ -30,6 +31,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Refactored error handling.
 * Added support for multiple data providers
 * Removed project creator for Jira.
+* Added client view.
+* Added account view.
+* Added leantime support for projects and project sync.
 
 
 * RELEASE NOTES: